# Performance improvements related to #32536 