Check out [related openpilot PR](https://github.com/commaai/openpilot/pull/35183) for more details

**Description**

Below, I will outline opendbc-specific changes I made to improve the test's performance:

1. Precompute `KF1D` parameters for `CarStateBase.__init__()`: total time execution time went down from ~12.2s to ~9.8s
    - Rationale: `get_kalman_gain()` is non-probabilistic, has no side effects, and always takes a set of constant parameters
2. Ford-specific optimization: introduce cache for `CANDefine` objects created during `CarState` construction, which decreased the per-car execution time from 0.8s (given all the above optimizations) to ~0.54s